updated welcome page
