+++ public/index.html
@@ -27,6 +27,6 @@
     <div id="root"></div>
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-    <script src="./js/index.js"></script>
+    <!-- <script src="./js/frontPage.js"></script> -->
   </body>
 </html>
