--- public/js/index.js
@@ -1,31 +1,0 @@
-  var lFollowX = 0,
-    lFollowY = 0,
-    x = 0,
-    y = 0,
-    friction = 1 / 20;
-
-  function moveBackground() {
-    x += (lFollowX - x) * friction;
-    y += (lFollowY - y) * friction;
-    
-    var translate = 'translate(' + x + 'px, ' + y + 'px) scale(1.1)';
-
-    $('.bg').css({
-      '-webit-transform': translate,
-      '-moz-transform': translate,
-      'transform': translate
-    });
-
-    window.requestAnimationFrame(moveBackground);
-  }
-
-  $(window).on('mousemove click', function(e) {
-
-    var lMouseX = Math.max(-100, Math.min(40, $(window).width() / 2 - e.clientX));
-    var lMouseY = Math.max(-100, Math.min(40, $(window).height() / 2 - e.clientY));
-    lFollowX = (20 * lMouseX) / 100; // 100 : 12 = lMouxeX : lFollow
-    lFollowY = (10 * lMouseY) / 100;
-
-  });
-
-  moveBackground();
