--- src/components/App/App.css
@@ -1,37 +1,0 @@
-h1:hover{
-  color:#000;
-  -webkit-transition: all .3s ease-in-out;
-  -moz-transition: all .3s ease-in-out;
-  transition: all .3s ease-in-out;
-}
-.wrap {
-  height: 100%;
-  position: relative;
-  overflow: hidden;
-}
-.image-wrap{
-  height: 100%;
-  position: relative;
-}
-.top-image{
-  width: auto;
-  height: 80%;
-  margin-left: auto;
-  margin-right: auto;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-}
-.bg {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: -1;
-  opacity: 0.6;
-  background: url('https://s3-us-west-2.amazonaws.com/kendrickdamnproject/kendrick+lamar+damn/welcome.jpg') no-repeat center center;
-  background-size: cover;
-  transform: scale(1.1);
-}
