+++ src/components/Welcome/index.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
-import './app.css';
+import FrontPage from '../../../public/js/frontPage.js';
+import './welcome.css';
 
-class App extends Component {
+class Welcome extends Component {
 
   render() {
 
@@ -10,11 +11,13 @@ class App extends Component {
         <div className="bg"></div>
         <div className="image-wrap">
           <img className="top-image" src="https://s3-us-west-2.amazonaws.com/kendrickdamnproject/kendrick+lamar+damn/welcome-top.png" alt="#" />
-        <a href="/home"><h1>damn.</h1></a>
+          <a href="/home"><h1>damn.</h1></a>
         </div>
+        // eslint-disable-next-line
+        <script>FrontPage</script>
       </div>
     );
   }
 }
 
-export default App;
+export default Welcome;
