+++ src/components/Welcome/welcome.css
@@ -1,0 +1,60 @@
+h1 {
+  margin: 0;
+  padding: 0;
+  position: absolute;
+  top: 20%;
+  left: 50%;
+  color: #fefefe;
+  font-size: 16vmin;
+  text-align: center;
+  text-transform: uppercase;
+  transform: translate3d(-50%, -50%, 0);
+}
+h1:hover{
+  color:#000;
+  -webkit-transition: all .3s ease-in-out;
+  -moz-transition: all .3s ease-in-out;
+  transition: all .3s ease-in-out;
+}
+.wrap {
+  height: 100%;
+  position: relative;
+  overflow: hidden;
+}
+.image-wrap{
+  height: 100%;
+  position: relative;
+}
+.top-image{
+  width: auto;
+  height: 80%;
+  margin-left: auto;
+  margin-right: auto;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+}
+.bg {
+  width: 100%;
+  height: 92%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: -1;
+  background: linear-gradient(rgba(255,255,255,0.4),rgba(255,255,255,0.4)), url('https://s3-us-west-2.amazonaws.com/kendrickdamnproject/kendrick+lamar+damn/welcome.jpg') no-repeat center center;
+  background-size: cover;
+  transform: scale(1.1);
+}
+
+@media only screen and (max-width: 768px) {
+  h1 {
+    color: #000;
+    top: 35%;
+    padding: 0 15px;
+    border: 3px solid #fefefe;
+  }
+  .top-image {
+    display:none;
+  }
+}
