+++ src/index.css
@@ -15,15 +15,3 @@ body {
   padding: 20px;
   font-family: 'Libre Baskerville', serif;
 }
-h1 {
-  margin: 0;
-  padding: 0;
-  position: absolute;
-  bottom: 20%;
-  left: 50%;
-  color: #fefefe;
-  font-size: 16vmin;
-  text-align: center;
-  text-transform: uppercase;
-  transform: translate3d(-50%, -50%, 0);
-}
