+++ src/routes.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Router, Route } from 'react-router';
-import App from './components/App';
+import Welcome from './components/Welcome';
 import Home from './components/Home';
 
 export default (
 	<Router>
-		<Route path={'/'} component={App} />
+		<Route path={'/'} component={Welcome} />
 		<Route path={'home'} component={Home} />
 	</Router>
 )
